[feat] complete Stage 1
Initial program that takes a string as input for a tic-tac-toe board. Print the board out in a formatted fashion, then get input from the player as to where to play. After making sure it's a valid move, update the board with 'X' or 'O' as appropriate. Print the board out again followed by a message as to the boards state - i.e. if it's a draw, not finished or there's a winner.